Move `alloc_impls` module to dedicated file
Per request of @HeroicKatora in PR #80
